// src/app/model/admService.ts
export class admService  {  



                serviceId: null;
                serviceDescription: null;
                userId: null;
                contDrAcctNo: null;
                contDrAcctType: null;
                contDrAcctNarr: null;
                contCrAcctNo: null;
                contCrAcctType: null;
                contCrAcctNarr: null;
                defaultDept: number;
                dateCreated: null;
                status: null;
                incomeAcctNo: null;
                incomeAcctType: null;
                channel: null;
                Frequency: null;
                Sequence: null;
                RefPrefix: null;
                TransactionDate: null;
                CustAcctDrTC: null;
                GLAcctDrTC: null;
                CustAcctCrTC: null;
                GLAcctCrTC: null;
                ReqAmortSched: null;
}
        