// src/app/model/overDraft.model.ts
export class OverDraft {


    itbId: number;
    serviceId: number;
    origDeptId: string;
    referenceNo: string;
    acctType: string;
    acctNo: string;
    acctName: string;
    ccyCode: string;
    availBal: string;
    acctSic: string;
    acctStatus: string;
    branchNo: string;
    acctProductCode: string;
    acctCustNo: string;
    odExpiryDate: string;
    drIntRate: string;
    odlimit: string;
    odOption: string;
    unapprovedOdRate: string;
    processingDeptId: string;
    odType: string;
    startDate: string;
    tenor: string;
    tenorPeriod: string;
    newExpiryDate: string;
    approvedLimit: string;
    approvedOdRate: string;
    templateType: string;
    serviceCcy: string;
    rsmId: number;
    incomeBranch: string;
    rejected: string;
    rejectedDate: string;
    parentId: string;
    serviceStatus: string;
    transactionDate: string;
    valueDate: string;
    dateCreated: string;
    status: string;
    userId: string;
    supervisorId: string;
    approvalDate: string;
    originatingBranchId: string;
    amortDate: string;
    rejectedIds: string;
    rejectedBy: string;
    industrySector: string;
    oDTenor: string;
    oDTimeBasis: string;
    batchNo: string;
    dismissedBy: string;
    dismissedDate: string;
    endDate: string;
    select: boolean;
    userName: string;
    branchName: string;
    createdBy: string;
    userProcessingDept: string;
    originBranch   : string;

}